+++ src/Tizen.Applications.Common/Interop/Interop.Event.cs
@@ -1,5 +1,5 @@
 ﻿/*
- * Copyright (c) 2016 Samsung Electronics Co., Ltd All Rights Reserved
+ * Copyright (c) 2018 Samsung Electronics Co., Ltd All Rights Reserved
  *
  * Licensed under the Apache License, Version 2.0 (the License);
  * you may not use this file except in compliance with the License.
@@ -20,7 +20,7 @@ using Tizen.Applications;
 
 internal static partial class Interop
 {
-    internal static partial class EventSystem
+    internal static partial class Event
     {
         internal enum ErrorCode
         {
